+++ content/themes/air-blocks/js/src/front-end.js
@@ -3,7 +3,7 @@
  * @Author: Roni Laukkarinen
  * @Date:   2021-11-18 15:12:35
  * @Last Modified by:   Roni Laukkarinen
- * @Last Modified time: 2022-06-07 17:00:47
+ * @Last Modified time: 2022-12-05 18:48:49
  */
 
 // Import modules (comment to disable)
@@ -15,10 +15,9 @@ import initAnchors from './modules/anchors';
 import backToTop from './modules/top';
 import initAirBlocksDemoFunctionality from './modules/air-blocks-demo';
 import initCarousels from './modules/carousels';
-import 'what-input';
+import init100vh from './modules/100vh';
+import initAccordions from './modules/accordion';
 import './modules/navigation';
-import './modules/100vh';
-import './modules/accordion';
 
 // Define Javascript is active by changing the body class
 document.body.classList.remove('no-js');
@@ -32,6 +31,8 @@ document.addEventListener('DOMContentLoaded', () => {
   styleExternalLinks();
   initExternalLinkLabels();
   getChildAltText();
+  initAccordions();
+  init100vh();
 
   // Fit video embeds to container
   reframe('.wp-has-aspect-ratio iframe');